requirements-dev.txt
-r requirements.txt
pytest-django==4.4.0
pytest==6.2.4
factory-boy==3.2.0
